fix: resolve Wave 2 compilation errors (round 2)
Fix remaining Kotlin compilation errors across 18 files:
- GroupDef: rename isRepeat field to mIsRepeat to avoid property/method clash
- GroupDef: add !! for nullable getBind() in ExtWrapTagged
- DateData: rename init field to mInit to avoid property/method clash
- ExternalDataInstanceSource: rename useCaseTemplate field to mUseCaseTemplate
- ExternalDataInstanceSource: add @Suppress for unchecked cast in buildVirtual
- DataInstance: fix getCurrentQueryContext() call (not a property)
- DataInstance: add @Suppress for WrapWithContext unchecked cast
- DataInstance: guard referenceCache.register with null check
- DataInstance: add @Suppress for getAttribute unchecked cast
- ItemSetUtils: fix getCurrentQueryContext() calls (not properties)
- Selection: use getIndex() getter instead of private field access
- TreeReferenceLevel: fix nullable name.hashCode() with ?: 0
- TreeReferenceLevel: fix attachCacheTable parameter shadowing companion field
- TreeReference: rename hashCode field to cachedHashCode to avoid method shadow
- FormIndex: add !! for nullable diff() return in constructor call
- EvaluationTraceReduction: fix Long/Int division with .toLong()
- QuestionString: add setTextId() method for QuestionDef.setTextID usage
- ItemsetBinding: add !! for nullable refs in writeExternal
- Triggerable: fix nullable expr?.hashCode() in hashCode()
- SendAction: add !! for nullable name and getString() return values
- SendAction: use getActionController()!! instead of property access
- TreeElement: fix PrototypeFactory? to non-null in private methods
- TreeUtilities: add null check for name before batch child fetch
- RestoreUtils: handle nullable contextualize() return value
- InstrumentationUtils: collect traces to list before iterating

+++ src/main/java/org/javarosa/core/model/trace/EvaluationTraceReduction.kt
@@ -89,7 +89,7 @@ class EvaluationTraceReduction(trace: EvaluationTrace) : EvaluationTrace(trace.g
     override fun getProfileReport(): String {
         var response = "{\n"
         response += "    time: " + getRuntimeCount(getRuntimeInNanoseconds()) + "\n"
-        response += "    time/call: " + getRuntimeCount(getRuntimeInNanoseconds() / countExecuted) + "\n"
+        response += "    time/call: " + getRuntimeCount(getRuntimeInNanoseconds() / countExecuted.toLong()) + "\n"
         var valueResponseCount = 0
         val totalRecords = valueMap.size
         for (key in valueMap.keys) {
